% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_rates_lrt-class.R
\name{BetaViolinPlot}
\alias{BetaViolinPlot}
\title{BetaViolinPlot}
\usage{
BetaViolinPlot(object, file = NULL, width = 8, height = 6, dpi = 300)
}
\arguments{
\item{object}{an \code{\linkS4class{TranscriptionRatesLRT}} object}

\item{file}{the path to a file to save the plot to}

\item{width}{the width of the plot in inches}

\item{height}{the height of the plot in inches}

\item{dpi}{the resolution of the plot in dpi}
}
\value{
an \code{\link{ggplot2}} object
}
\description{
Plot a violin plot comparing the beta values between two conditions.
}
\examples{
# Create an ExperimentTranscriptionRates object
load("inst/extdata/granges_for_read_counting_DLD1_chr21.RData")
transcriptionRates1 <- estimateTranscriptionRates(
    "inst/extdata/PROseq-DLD1-aoi-NELFC_Auxin_Ctrl-SE_plus_chr21.bw",
    bigwigMinus = 
     "inst/extdata/PROseq-DLD1-aoi-NELFC_Auxin_Ctrl-SE_minus_chr21.bw",
    pauseRegions = bw_pause_filtered,
    geneBodyRegions = bw_gb_filtered,
    name = "Control"
)
transcriptionRates2 <- estimateTranscriptionRates(
    "inst/extdata/PROseq-DLD1-aoi-NELFC_Auxin_Ctrl-SE_plus_chr21.bw",
    bigwigMinus = 
     "inst/extdata/PROseq-DLD1-aoi-NELFC_Auxin_Ctrl-SE_minus_chr21.bw",
    pauseRegions = bw_pause_filtered,
    geneBodyRegions = bw_gb_filtered,
    name = "Treated"
)
spikeInFile <- "inst/extdata/spikein_scaling_factor.csv"
lrts <- likelihoodRatioTest(transcriptionRates1, transcriptionRates2,
spikeInFile)
BetaViolinPlot(lrts, file="boxplot.png")

}
