% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenomic_features-class.R
\docType{class}
\name{EpigenomicFeatures-class}
\alias{EpigenomicFeatures-class}
\alias{preprocessEpigenomicFeatures}
\title{Constructor for EpigenomicFeatures object}
\usage{
preprocessEpigenomicFeatures(
  bigwigPlus,
  bigwigMinus,
  geneBodyRegions,
  histoneInputFolder,
  ctcfBigwig,
  wgbsBedfile,
  spliceJunctionsBedfile,
  stemLoopsBedfile,
  repeatsBedfile,
  chromosomes = c("22")
)
}
\arguments{
\item{bigwigPlus}{The path to a bigwig file for plus strand QC recording
PRO-seq read counts. This can be generated with the proseq2.0 pipeline.}

\item{bigwigMinus}{The path to a bigwig file for minus strand QC recording #' PRO-seq read counts. This can be generated with the proseq2.0 pipeline.}

\item{geneBodyRegions}{A \code{\link[GenomicRanges]{GRanges-class}} object
that holds the gene body regions coordinates used for gene coordinate
binning for feature smoothing}

\item{histoneInputFolder}{The path to a folder containing the histone bigwig #' input files}

\item{ctcfBigwig}{The path to a bigwig file containing the CTCF signal}

\item{wgbsBedfile}{The path to a bed file containing the WGBS signal}

\item{spliceJunctionsBedfile}{The path to a bed file containing the splice
junction coordinates and strand information}

\item{stemLoopsBedfile}{The path to a bed file containing the stem-loop
coordinates and strand information}

\item{repeatsBedfile}{The path to a bed file containing the repeat
coordinates for low complexity regions}

\item{chromosomes}{a character vector of the chromosomes to process
(default: # chr22). NULL means all chromosomes.}
}
\value{
an \code{\link{EpigenomicFeatures-class}} object
}
\description{
Class \code{EpigenomicFeatures}...

Preprocessing
}
\section{Slots}{

\describe{
\item{\code{bigwigPlus}}{a path to bigwig for plus strand QC recording PRO-seq read
counts. This can be generated with the proseq2.0 pipeline.}

\item{\code{bigwigMinus}}{a path to bigwig for minus strand QC recording PRO-seq read
counts. This can be generated with the proseq2.0 pipeline.}

\item{\code{geneBodyRegions}}{a \code{\link[GenomicRanges]{GRanges-class}} object
that holds the gene body regions coordinates used for gene coordinate
binning for feature smoothing}

\item{\code{histoneInputFolder}}{a path to folder containing the histone bigwig
input files}

\item{\code{chromosomes}}{a character vector of the chromosomes to process (default: # chr22). NULL means all chromosomes.}

\item{\code{repeatsBedfile}}{a path to a bed file containing the repeat coordinates #' for low complexity regions}

\item{\code{ctcfBigwig}}{a path to a bigwig file containing the CTCF signal}

\item{\code{wgbsBedfile}}{a path to a bed file containing the WGBS signal}

\item{\code{spliceJunctionsBedfile}}{a path to a bed file containing the splice
junction coordinates and strand information}

\item{\code{stemLoopsBedfile}}{a path to a bed file containing the stem-loop
coordinates and strand information}
}}

