% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{estimate_experiment_transcription_rates}
\alias{estimate_experiment_transcription_rates}
\title{estimate_experiment_transcription_rates}
\usage{
estimate_experiment_transcription_rates(
  bigwig_plus,
  bigwig_minus,
  pause_regions,
  gene_body_regions,
  gene_name_column,
  steric_hindrance = FALSE,
  omega_scale = NULL
)
}
\arguments{
\item{bigwig_plus}{the path to a bigwig file from the plus strand recording PRO-seq read counts}

\item{bigwig_minus}{the path to a bigwig file from the minus strand recording PRO-seq read counts}

\item{pause_regions}{a \link[GenomicRanges]{GRanges-class} object that must contain a gene_id
metadata column}

\item{gene_body_regions}{a \link[GenomicRanges]{GRanges-class} object that must contain a gene_id
metadata column}

\item{gene_name_column}{a string that indicates which column in the GRanges object contains the
gene names}

\item{steric_hindrance}{a logical value to determine whether to infer landing-pad occupancy or not.
Defaults to FALSE.}

\item{omega_scale}{a numeric value for scaling omega. Defaults to NULL.}
}
\value{
an \code{\link{experiment_transcription_rates-class}} object
}
\description{
Estimates the transcription rates from experimental data and contructs an object that holds
these rates
}
