% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{estimateExperimentTranscriptionRates}
\alias{estimateExperimentTranscriptionRates}
\title{estimateExperimentTranscriptionRates}
\usage{
estimateExperimentTranscriptionRates(
  bigwigPlus,
  bigwigMinus,
  pauseRegions,
  geneBodyRegions,
  geneNameColumn = "gene_id",
  stericHindrance = FALSE,
  omegaScale = NULL
)
}
\arguments{
\item{bigwigPlus}{the path to a bigwig file from the plus strand recording
PRO-seq read counts}

\item{bigwigMinus}{the path to a bigwig file from the minus strand recording
PRO-seq read counts}

\item{pauseRegions}{a \link[GenomicRanges]{GRanges-class} object that must
contain a geneId}

\item{geneBodyRegions}{a \link[GenomicRanges]{GRanges-class} object that
must contain a geneId}

\item{geneNameColumn}{a string that indicates which column in the GRanges
represents gene names information. Defaults to "gene_id"}

\item{stericHindrance}{a logical value to determine whether to infer
landing-pad occupancy or not. Defaults to FALSE.}

\item{omegaScale}{a numeric value for scaling omega. Defaults to NULL.}
}
\value{
an \code{\link{experimentTranscriptionRates-class}} object
}
\description{
Estimates the transcription rates, such as initiation, pause-release rates
and landing pad occupancy, from experimental data, such as nascent RNA
sequencing read counts and genomic coordinates, and contructs an object
that holds these rates
}
