% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\name{estimateSimulationTranscriptionRates}
\alias{estimateSimulationTranscriptionRates}
\title{estimateSimulationTranscriptionRates}
\usage{
estimateSimulationTranscriptionRates(simpol, stericHindrance = FALSE)
}
\arguments{
\item{simpol}{a \code{\link{simulatePolymerase-class}} object}

\item{stericHindrance}{a logical value to determine whether to infer
landing-pad occupancy or not.
Defaults to FALSE.}
}
\value{
an \code{\link{simulationTranscriptionRates-class}} object
}
\description{
Estimates the transcription rates, such as initiation, pause-release rates
and landing pad occupancy, from simulated data, such as nascent RNA
sequencing read counts and genomic coordinates, and contructs an object
that holds these rates
}
