% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{estimateTranscriptionRates}
\alias{estimateTranscriptionRates}
\title{Generic function for estimating transcription rates}
\usage{
estimateTranscriptionRates(x, ...)
}
\arguments{
\item{x}{The input data (either a SimulatePolymerase object or bigwig files)}

\item{...}{Additional arguments passed to the specific methods}
}
\value{
An object containing estimated transcription rates
}
\description{
Generic function that estimates transcription rates from either simulation
data (SimulatePolymerase object) or experimental data (bigwig files and
genomic regions).
}
