% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\docType{class}
\name{experiment_transcription_rates-class}
\alias{experiment_transcription_rates-class}
\title{Class experiment_transcription_rates}
\description{
Class \code{experiment_transcription_rates} has read counts, pause and gene body genomic region
coordinates, steric hindrance and omega scale factor parameters used to estimate the transcription
rates
}
\section{Slots}{

\describe{
\item{\code{counts}}{a \code{data.frame} with five columns gene_id, summarized_pause_counts, pause_length
summarized_gb_counts, gb_length}

\item{\code{bigwig_plus}}{a path to bigwig for plus strand}

\item{\code{bigwig_minus}}{a path to bigwig for minus strand}

\item{\code{pause_regions}}{a \code{\link[GenomicRanges]{GRanges-class}} that holds
all the pause region coordinates}

\item{\code{gene_body_regions}}{a \code{\link[GenomicRanges]{GRanges-class}} that holds
all the gene body region coordinates}

\item{\code{gene_name_column}}{a string for the gene name column in the GRanges objects}

\item{\code{steric_hindrance}}{a logical value representing whether landing-pad occupancy was inferred when
estimating the rates}

\item{\code{omega_scale}}{a numeric value for the scale factor used to calculate omega}

\item{\code{rates}}{a \code{tbl_df} containing estimated transcription rates such as chi estimates,
beta_org estimates from the initial model, beta_adp estimates from the model with varying
pause sites, fk_mean giving the mean position of pause sites, fk_var for variance of pause sites,
phi estimates for landing-pad occupancy, omega_zeta for the effective initiation rate, beta_zeta
for the pause-escape rate, alpha_zeta for the potential initiation rate, and likelihoods}
}}

