% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{experiment_transcription_rates_valid}
\alias{experiment_transcription_rates_valid}
\title{Check validity of experiment_transcription_rates object}
\usage{
experiment_transcription_rates_valid(object)
}
\arguments{
\item{object}{A \code{experiment_transcription_rates} object}
}
\value{
TRUE if valid, else errors
}
\description{
Validates an experiment_transcription_rates object
}
\keyword{internal}
