% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{experimentTranscriptionRatesValid}
\alias{experimentTranscriptionRatesValid}
\title{Check validity of experimentTranscriptionRates object}
\usage{
experimentTranscriptionRatesValid(object)
}
\arguments{
\item{object}{A \code{experimentTranscriptionRates} object}
}
\value{
TRUE if valid, else errors
}
\description{
Validates an experimentTranscriptionRates object
}
\keyword{internal}
