% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{get_parameters}
\alias{get_parameters}
\title{Get all simulation parameters}
\usage{
get_parameters(object)
}
\arguments{
\item{object}{A simulate_polymerase object}
}
\value{
A list containing all simulation parameters
}
\description{
Get all simulation parameters
}
