% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{get_pause_sites_df}
\alias{get_pause_sites_df}
\title{Get pause sites as a data frame}
\usage{
get_pause_sites_df(object)
}
\arguments{
\item{object}{A simulate_polymerase object}
}
\value{
A data frame with cell numbers and their pause sites
}
\description{
Get pause sites as a data frame
}
