% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{get_polymerase_counts_df}
\alias{get_polymerase_counts_df}
\title{Get combined cells data as a data frame}
\usage{
get_polymerase_counts_df(object)
}
\arguments{
\item{object}{A simulate_polymerase object}
}
\value{
A data frame with positions and their polymerase counts
}
\description{
Get combined cells data as a data frame
}
