% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{get_position_df}
\alias{get_position_df}
\title{Get position matrix as a tidy data frame}
\usage{
get_position_df(object)
}
\arguments{
\item{object}{A simulate_polymerase object}
}
\value{
A data frame with cell, position, and polymerase presence
}
\description{
Get position matrix as a tidy data frame
}
