% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{get_probability_df}
\alias{get_probability_df}
\title{Get probability vector as a data frame}
\usage{
get_probability_df(object)
}
\arguments{
\item{object}{A simulate_polymerase object}
}
\value{
A data frame with positions and their transition probabilities
}
\description{
Get probability vector as a data frame
}
