% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{getCounts}
\alias{getCounts}
\title{Get read counts}
\usage{
getCounts(object)
}
\arguments{
\item{object}{An experimentTranscriptionRates object}
}
\value{
A data.frame containing the read counts
}
\description{
Get read counts
}
