% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{getParameters}
\alias{getParameters}
\title{Get all simulation parameters}
\usage{
getParameters(object)
}
\arguments{
\item{object}{A simulatePolymerase object}
}
\value{
A list containing all simulation parameters
}
\description{
Get all simulation parameters
}
