% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{getPauseSitesDf}
\alias{getPauseSitesDf}
\title{Get pause sites as a data frame}
\usage{
getPauseSitesDf(object)
}
\arguments{
\item{object}{A simulatePolymerase object}
}
\value{
A data frame with cell numbers and their pause sites
}
\description{
Get pause sites as a data frame
}
