% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{getPolymeraseCountsDf}
\alias{getPolymeraseCountsDf}
\title{Get combined cells data as a data frame}
\usage{
getPolymeraseCountsDf(object)
}
\arguments{
\item{object}{A simulatePolymerase object}
}
\value{
A data frame with positions and their polymerase counts
}
\description{
Get combined cells data as a data frame
}
