% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{getPositionDf}
\alias{getPositionDf}
\title{Get position matrix as a tidy data frame}
\usage{
getPositionDf(object)
}
\arguments{
\item{object}{A simulatePolymerase object}
}
\value{
A data frame with cell, position, and polymerase presence
}
\description{
Get position matrix as a tidy data frame
}
