% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{getProbabilityDf}
\alias{getProbabilityDf}
\title{Get probability vector as a data frame}
\usage{
getProbabilityDf(object)
}
\arguments{
\item{object}{A simulatePolymerase object}
}
\value{
A data frame with positions and their transition probabilities
}
\description{
Get probability vector as a data frame
}
