% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{getRates}
\alias{getRates}
\title{Get transcription rates}
\usage{
getRates(object)
}
\arguments{
\item{object}{An experiment_transcription_rates object}
}
\value{
A tibble containing the transcription rates
}
\description{
Get transcription rates
}
