% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_rates_lrt-class.R
\name{getSdGroupCutoff}
\alias{getSdGroupCutoff}
\title{Get SD Group Cutoff}
\usage{
getSdGroupCutoff(df, value_col, group_col)
}
\arguments{
\item{df}{a \code{tbl_df} or \code{data.frame} containing \code{value_col}
and \code{group_col}}

\item{value_col}{the name of the column with fkSD values}

\item{group_col}{the name of the column with "Sharp"/"Broad" group labels}
}
\value{
a numeric cutoff value
}
\description{
Computes the midpoint cutoff between the maximum fkSD value among "Sharp"
genes and the minimum fkSD value among "Broad" genes for a given fkSD
column and its associated sdGroup column. Ported from
\code{get_sdgroup_cutoff} in \code{1_2_LRT_viz.R}.
}
