% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_rates_lrts-class.R
\name{likelihood_ratio_test}
\alias{likelihood_ratio_test}
\title{likelihood_ratio_test}
\usage{
likelihood_ratio_test(exp_data1, exp_data2, sc)
}
\arguments{
\item{exp_data1}{an \code{\link{experiment_transcription_rates-class}} object}

\item{exp_data2}{an \code{\link{experiment_transcription_rates-class}} object}

\item{sc}{path to a csv file containing scale factors based on total or
spike-in reads}
}
\value{
a \code{\link{likelihood_ratio_test-class}} object
}
\description{
DESCRIPTION
}
