% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_rates_lrt-class.R
\name{likelihoodRatioTest}
\alias{likelihoodRatioTest}
\title{likelihoodRatioTest}
\usage{
likelihoodRatioTest(expData1, expData2, spikeInScalingFactor)
}
\arguments{
\item{expData1}{an \code{\link{experimentTranscriptionRates-class}} object}

\item{expData2}{an \code{\link{experimentTranscriptionRates-class}} object}

\item{spikeInScalingFactor}{path to a csv file containing scale factorsbased
on total or spike-in reads

Note: Gene body length assumed to be the same between conditions}
}
\value{
a \code{\link{likelihoodRatioTest-class}} object
}
\description{
DESCRIPTION
}
\examples{
expData1 <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
expData2 <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
spikeInScalingFactor <- "path/to/scale.csv"
lrts <- likelihoodRatioTest(expData1, expData2, spikeInScalingFactor)
# Print the likelihood ratio test object
print(lrts)
}
