% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{parameters}
\alias{parameters}
\alias{parameters,SimulatePolymerase-method}
\title{Get all simulation parameters}
\usage{
parameters(object)

\S4method{parameters}{SimulatePolymerase}(object)
}
\arguments{
\item{object}{A SimulatePolymerase object}
}
\value{
A list containing all simulation parameters including k, ksd, kMin,
kMax, geneLen, alpha, beta, zeta, zetaSd, zetaMin, zetaMax, zetaVec, cellNum,
polSize, addSpace, time, and stepsToRecord
}
\description{
Get all simulation parameters
}
\examples{
# Create a SimulatePolymerase object
sim <- SimulatePolymerase(
    k=50, ksd=25, kMin=17, kMax=200, geneLen=1950,
    alpha=1, beta=1, zeta=2000, zetaSd=1000, zetaMin=1500, zetaMax=2500,
    zetaVec=NULL, cellNum=1000, polSize=33, addSpace=17, time=1, 
    stepsToRecord=1)
# Get parameters
parameters <- parameters(sim)
# Print the parameters
print(parameters)
}
