% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_estimation_helpers.R
\name{pause_escape_EM}
\alias{pause_escape_EM}
\title{Pause Escape Expectation Maximization}
\usage{
pause_escape_EM(
  fk_int,
  Xk,
  kmin,
  kmax,
  beta_int,
  chi_hat,
  max_itr = 100,
  tor = 0.001
)
}
\arguments{
\item{fk_int}{a list of the initial pause site values}

\item{Xk}{a numeric vector of read counts on each position within the pause peak}

\item{kmin}{an integer for lower bound of pause sites}

\item{kmax}{an integer for upper bound of pause sites}

\item{beta_int}{a list of initialized beta estimates}

\item{chi_hat}{a numeric for read count chi estimate}

\item{max_itr}{an integer for the maximum iterations. Default is 100.}

\item{tor}{Tolerance value to determine when to stop iterating. Default is 1e-3}
}
\value{
A list of transcription rates including beta, Yk, fk, fk_mean, fk_var, betas,
likelihoods and phi
}
\description{
Estimate transcription rates with varying pause sites.
}
