% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_estimation_helpers.R
\name{pauseEscapeEM}
\alias{pauseEscapeEM}
\title{Pause Escape Expectation Maximization}
\usage{
pauseEscapeEM(
  fkInt,
  Xk,
  kmin,
  kmax,
  betaInt,
  chiHat,
  maxItr = 100,
  tor = 0.001
)
}
\arguments{
\item{fkInt}{a list of the initial pause site values}

\item{Xk}{a numeric vector of read counts on each position within the
pause peak}

\item{kmin}{an integer for lower bound of pause sites}

\item{kmax}{an integer for upper bound of pause sites}

\item{betaInt}{a list of initialized beta estimates}

\item{chiHat}{a numeric for read count chi estimate}

\item{maxItr}{an integer for the maximum iterations. Default is 100.}

\item{tor}{Tolerance value to determine when to stop iterating.
Default is 1e-3}
}
\value{
A list of transcription rates including beta, Yk, fk, fk_mean,
fk_var, betas,
likelihoods and phi
}
\description{
Estimate transcription rates with varying pause sites.
}
