% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\name{plot_rnap_counts}
\alias{plot_rnap_counts}
\title{Plot RNAP counts}
\usage{
plot_rnap_counts(object, file = NULL, width = 8, height = 6)
}
\arguments{
\item{object}{A simulation_transcription_rates object}

\item{file}{Optional file path to save the plot}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\value{
A ggplot object showing the RNAP counts
}
\description{
Plot RNAP counts
}
