% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\name{plotPauseSiteDistribution}
\alias{plotPauseSiteDistribution}
\title{Plot pause site distribution}
\usage{
plotPauseSiteDistribution(object, file = NULL, width = 8, height = 6)
}
\arguments{
\item{object}{A simulationTranscriptionRates object}

\item{file}{Optional file path to save the plot}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\value{
A ggplot object showing the pause site distribution
}
\description{
Plot pause site distribution
}
