% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{plotPolymeraseDistribution}
\alias{plotPolymeraseDistribution}
\title{Plot polymerase distribution}
\usage{
plotPolymeraseDistribution(object, file = NULL, width = 8, height = 6)
}
\arguments{
\item{object}{A simulatePolymerase object}

\item{file}{Optional file path to save the plot}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\value{
A ggplot object showing the distribution of polymerases across the
gene
}
\description{
Plot polymerase distribution
}
