% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{sample_gene_body_avg_read_density}
\alias{sample_gene_body_avg_read_density}
\title{Sample read counts from a simulate_polymerase object}
\usage{
sample_gene_body_avg_read_density(object, read_density = 0.0489)
}
\arguments{
\item{object}{A simulate_polymerase object}

\item{read_density}{A numeric value for the read density within gene body in
\emph{Dukler et al.} (2017) for genes with median expression (i.e., 0.0489).}
}
\value{
The read count per nucleotide value
}
\description{
Sample read counts from a simulate_polymerase object
}
