% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{sampleGeneBodyAvgReadDensity}
\alias{sampleGeneBodyAvgReadDensity}
\title{Sample read counts from a simulatePolymerase object}
\usage{
sampleGeneBodyAvgReadDensity(object, readDensity = 0.0489)
}
\arguments{
\item{object}{A simulatePolymerase object}

\item{readDensity}{A numeric value for the read density within gene body in
\emph{Dukler et al.} (2017) for genes with median expression (i.e., 0.0489).}
}
\value{
The read count per nucleotide value
}
\description{
Sample read counts from a simulatePolymerase object
}
