% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{save_data_frames}
\alias{save_data_frames}
\title{Save all data frames to CSV files}
\usage{
save_data_frames(object, dir = "results")
}
\arguments{
\item{object}{A simulate_polymerase object}

\item{dir}{Directory to save the files (default: "results")}
}
\description{
Save all data frames to CSV files
}
