% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{save_plots}
\alias{save_plots}
\title{Save all plots to files}
\usage{
save_plots(object, dir = "results", width = 8, height = 6)
}
\arguments{
\item{object}{A simulate_polymerase object}

\item{dir}{Directory to save the plots (default: "results")}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\description{
Save all plots to files
}
