% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{saveDataFrames}
\alias{saveDataFrames}
\title{Save all data frames to CSV files}
\usage{
saveDataFrames(object, dir = "results")
}
\arguments{
\item{object}{A simulatePolymerase object}

\item{dir}{Directory to save the files (default: "results")}
}
\value{
Outputs a CSV file with the data frames
}
\description{
Save all data frames to CSV files
}
