% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{savePlots}
\alias{savePlots}
\title{Save all plots to files}
\usage{
savePlots(object, dir = "results", width = 8, height = 6)
}
\arguments{
\item{object}{A simulatePolymerase object}

\item{dir}{Directory to save the plots (default: "results")}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\value{
Outputs directory where plots were saved to pdf files
}
\description{
Save all plots to files
}
\examples{
# Create a simulatePolymerase object
sim <- simulatePolymerase(
    k=50, ksd=25, kMin=17, kMax=200, geneLen=1950,
    alpha=1, beta=1, zeta=2000, zetaSd=1000, zetaMin=1500, zetaMax=2500,
    zetaVec=NULL, cellNum=1000, polSize=33, addSpace=17, time=1, 
    stepsToRecord=1)
# Save plots
savePlots(sim)
}
