% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{show,experiment_transcription_rates-method}
\alias{show,experiment_transcription_rates-method}
\title{Show method for experiment_transcription_rates objects}
\usage{
\S4method{show}{experiment_transcription_rates}(object)
}
\arguments{
\item{object}{An experiment_transcription_rates object}
}
\value{
NULL (invisibly)
}
\description{
Show method for experiment_transcription_rates objects
}
\examples{
# Create an experiment_transcription_rates object
exp_rates <- estimate_experiment_transcription_rates(
    bigwig_plus = "path/to/plus.bw",
    bigwig_minus = "path/to/minus.bw",
    pause_regions = GRanges("chr1:1-1000"),
    gene_body_regions = GRanges("chr1:1-2000"),
    gene_name_column = "gene_id"
)

# Show the object
show(exp_rates)
}
