% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{show,ExperimentTranscriptionRates-method}
\alias{show,ExperimentTranscriptionRates-method}
\title{Show method for ExperimentTranscriptionRates objects}
\usage{
\S4method{show}{ExperimentTranscriptionRates}(object)
}
\arguments{
\item{object}{An ExperimentTranscriptionRates object}
}
\value{
NULL (invisibly)
}
\description{
Show method for ExperimentTranscriptionRates objects
}
\examples{
# Create an ExperimentTranscriptionRates object
expRates <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)

# Show the object
show(expRates)
}
