% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\docType{class}
\name{simulate_polymerase-class}
\alias{simulate_polymerase-class}
\title{Class simulate_polymerase}
\description{
Class \code{simulate_polymerase} has
}
\section{Slots}{

\describe{
\item{\code{counts}}{a \code{data.frame} with five columns gene_id, summarized_pause_counts, pause_length
summarized_gb_counts, gb_length}
}}

