% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{simulate_polymerase}
\alias{simulate_polymerase}
\title{simulate_polymerase}
\usage{
simulate_polymerase(k, kSd, kMax)
}
\arguments{
\item{k}{a numeric (int) value for the mean of pause sites across cells. Defaults to 50bp.}

\item{kSd}{a numeric (double) value for std dev of pause sites across cells. Defaults to 0bp.}

\item{kMax}{a numeric (int) for the lower bound of allowed pause site. Defaults to 17bp.}
}
\value{
an \code{\link{experiment_transcription_rates-class}} object
}
\description{
Estimates
}
