% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\docType{class}
\name{simulation_transcription_rates-class}
\alias{simulation_transcription_rates-class}
\alias{get_simpol}
\alias{get_steric_hindrance}
\alias{get_trial}
\alias{get_chi}
\alias{get_beta_org}
\alias{get_beta_adp}
\alias{get_phi}
\alias{get_fk}
\alias{get_fk_mean}
\alias{get_fk_var}
\alias{get_flag}
\alias{get_rnap_n}
\alias{show,simulation_transcription_rates-method}
\title{Class simulation_transcription_rates}
\usage{
get_simpol(object)

get_steric_hindrance(object)

get_trial(object)

get_chi(object)

get_beta_org(object)

get_beta_adp(object)

get_phi(object)

get_fk(object)

get_fk_mean(object)

get_fk_var(object)

get_flag(object)

get_rnap_n(object)

\S4method{show}{simulation_transcription_rates}(object)
}
\description{
Class \code{simulation_transcription_rates} estimates the transcription
rates, such as initiation, pause-release rates and landing pad occupancy,
from simulated data, such as nascent RNA sequencing read counts and genomic
coordinates simulated from polymerase movement over a specified time period
calculated by simulator, and contructs an object that holds these rates.
}
