% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\name{simulation_transcription_rates_valid}
\alias{simulation_transcription_rates_valid}
\title{simulation_transcription_rates_valid}
\usage{
simulation_transcription_rates_valid(object)
}
\arguments{
\item{object}{a \code{simulation_transcription_rates} object}
}
\value{
TRUE if valid, else errors
}
\description{
Checks that a \code{simulation_transcription_rates} object is valid
}
