% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\docType{class}
\name{simulationTranscriptionRates-class}
\alias{simulationTranscriptionRates-class}
\alias{getSimpol}
\alias{getStericHindrance}
\alias{getTrial}
\alias{getChi}
\alias{getBetaOrg}
\alias{getBetaAdp}
\alias{getPhi}
\alias{getFk}
\alias{getFkMean}
\alias{getFkVar}
\alias{getFlag}
\alias{getRnapN}
\alias{show,simulationTranscriptionRates-method}
\title{Class simulationTranscriptionRates}
\usage{
getSimpol(object)

getStericHindrance(object)

getTrial(object)

getChi(object)

getBetaOrg(object)

getBetaAdp(object)

getPhi(object)

getFk(object)

getFkMean(object)

getFkVar(object)

getFlag(object)

getRnapN(object)

\S4method{show}{simulationTranscriptionRates}(object)
}
\description{
Class \code{simulationTranscriptionRates} estimates the transcription
rates, such as initiation, pause-release rates and landing pad occupancy,
from simulated data, such as nascent RNA sequencing read counts and genomic
coordinates simulated from polymerase movement over a specified time period
calculated by simulator, and contructs an object that holds these rates.
}
