% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\name{simulationTranscriptionRatesValid}
\alias{simulationTranscriptionRatesValid}
\title{simulationTranscriptionRatesValid}
\usage{
simulationTranscriptionRatesValid(object)
}
\arguments{
\item{object}{a \code{simulationTranscriptionRates} object}
}
\value{
TRUE if valid, else errors
}
\description{
Checks that a \code{simulationTranscriptionRates} object is valid
}
\keyword{internal}
