% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_estimation_helpers.R
\name{stericHindranceEM}
\alias{stericHindranceEM}
\title{Steric Hindrance Expectation-Maximization}
\usage{
stericHindranceEM(
  Xk,
  kmin,
  kmax,
  f1,
  f2,
  fkInt,
  betaInt,
  phiInt,
  chiHat,
  lambda,
  zeta,
  maxItr = 100,
  tor = 0.001
)
}
\arguments{
\item{Xk}{a numeric vector of read counts on each position within the pause
peak}

\item{kmin}{an integer for lower bound of pause sites}

\item{kmax}{an integer for upper bound of pause sites}

\item{f1}{a numeric}

\item{f2}{a numeric}

\item{fkInt}{a list of the initial pause site values}

\item{betaInt}{a list of initialized beta estimates}

\item{phiInt}{a numeric}

\item{chiHat}{a numeric for read count chi estimate}

\item{lambda}{a numeric for zeta scaled}

\item{zeta}{a numeric for elongation rate}

\item{maxItr}{an integer for the maximum iterations. Default is 100.}

\item{tor}{Tolerance value to determine when to stop iterating. Default is
1e-3}
}
\value{
A list of transcription rates including beta, Yk, fk, fk_mean,
fk_var, betas, likelihoods and phi
}
\description{
Estimate transcription rates with EM algorithm with varying pause sites and
steric hindrance.
Landing pad occupancy is inferred.
}
