% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_estimation_helpers.R
\name{summarise_bw}
\alias{summarise_bw}
\title{Summarize Bigwigs}
\usage{
summarise_bw(bw, grng, col_name)
}
\arguments{
\item{bw}{\code{\link[GenomicRanges]{GRanges-class}} object of read counts}

\item{grng}{\code{\link[GenomicRanges]{GRanges-class}} object of regions to
summarize read count}

\item{col_name}{string for the column name for the summarized read counts}
}
\value{
A \code{\link[GenomicRanges]{GenomicRanges-class}} object
}
\description{
Creates GRanges object with column of read counts summarized over regions of genes
}
