% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_estimation_helpers.R
\name{summariseBw}
\alias{summariseBw}
\title{Summarize Bigwigs}
\usage{
summariseBw(bw, grng, colName)
}
\arguments{
\item{bw}{\code{\link[GenomicRanges]{GRanges-class}} object of read counts}

\item{grng}{\code{\link[GenomicRanges]{GRanges-class}} object of regions to
summarize read count}

\item{colName}{string for the column name for the summarized read counts}
}
\value{
A \code{\link[GenomicRanges]{GenomicRanges-class}} object
}
\description{
Creates GRanges object with column of read counts summarized over regions of
genes
}
