% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_rates_lrts-class.R
\docType{class}
\name{transcription_rates_lrts-class}
\alias{transcription_rates_lrts-class}
\alias{exp_data1}
\alias{exp_data2}
\alias{sc}
\alias{likelihood_ratio_test}
\title{Class transcription_rates_lrts for comparing aspects of transcriptional
dynamics of the same TU under different conditions using Likelihood Ratio
Test Statistics}
\usage{
exp_data1(object)

exp_data2(object)

sc(object)

likelihood_ratio_test(exp_data1, exp_data2, sc)
}
\description{
Class \code{transcription_rates_lrts}
}
