% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_rates_lrts-class.R
\docType{class}
\name{transcriptionRatesLrts-class}
\alias{transcriptionRatesLrts-class}
\alias{expData1}
\alias{expData2}
\alias{sc}
\alias{likelihoodRatioTest}
\title{Class transcriptionRatesLrts for comparing aspects of transcriptional
dynamics of the same TU under different conditions using Likelihood Ratio
Test Statistics}
\usage{
expData1(object)

expData2(object)

sc(object)

likelihoodRatioTest(expData1, expData2, sc)
}
\description{
Class \code{transcriptionRatesLrts}
}
