% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\name{estimate_simulation_transcription_rates}
\alias{estimate_simulation_transcription_rates}
\title{estimate_simulation_transcription_rates}
\usage{
estimate_simulation_transcription_rates(simpol, steric_hindrance = FALSE)
}
\arguments{
\item{simpol}{a \code{\link{simulate_polymerase-class}} object}

\item{steric_hindrance}{a logical value to determine whether to infer
landing-pad occupancy or not.
Defaults to FALSE.}
}
\value{
an \code{\link{_transcription_rates-class}} object
}
\description{
Estimates the transcription rates, such as initiation, pause-release rates
and landing pad occupancy, from simulated data, such as nascent RNA
sequencing read counts and genomic coordinates, and contructs an object
that holds these rates
}
