% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\name{estimateSimulationTranscriptionRates}
\alias{estimateSimulationTranscriptionRates}
\title{estimateSimulationTranscriptionRates}
\usage{
estimateSimulationTranscriptionRates(simpol, stericHindrance = FALSE)
}
\arguments{
\item{simpol}{a \code{\link{simulatePolymerase-class}} object}

\item{stericHindrance}{a logical value to determine whether to infer
landing-pad occupancy or not.
Defaults to FALSE.}
}
\value{
an \code{\link{SimulationTranscriptionRates-class}} object
}
\description{
Estimates the transcription rates, such as initiation, pause-release rates
and landing pad occupancy, from simulated data, such as nascent RNA
sequencing read counts and genomic coordinates, and contructs an object
that holds these rates
}
\examples{
# Create a SimulatePolymerase object
sim <- simulatePolymerase(
    k=50, ksd=25, kMin=17, kMax=200, geneLen=1950,
    alpha=1, beta=1, zeta=2000, zetaSd=1000, zetaMin=1500, zetaMax=2500,
    zetaVec=NULL, cellNum=1000, polSize=33, addSpace=17, time=1, 
    stepsToRecord=1)
# Estimate transcription rates
estRates <- estimateSimulationTranscriptionRates(sim)
# Print the estimated rates
print(estRates)

}
