% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-transcription-rates-class.R
\name{estimateTranscriptionRates}
\alias{estimateTranscriptionRates}
\title{Generic function for estimating transcription rates}
\usage{
estimateTranscriptionRates(x, ...)
}
\arguments{
\item{x}{The input data (either a SimulatePolymerase object or bigwig files)}

\item{...}{Additional arguments passed to the specific methods}
}
\value{
An object containing estimated transcription rates
}
\description{
Generic function that estimates transcription rates from either simulation
data (SimulatePolymerase object) or experimental data (bigwig files and
genomic regions).
}
\examples{
load("inst/extdata/granges_for_read_counting_chr21_subset.RData")
expRates <- estimateTranscriptionRates(
    "inst/extdata/PROseq-K562-vihervaara-control-SE_plus_chr21_subset.bw",
    bigwigMinus = 
     "inst/extdata/PROseq-K562-vihervaara-control-SE_minus_chr21_subset.bw",
    pauseRegions = bw_pause_21_subset,
    geneBodyRegions = bw_gene_body_21_subset,
    stericHindrance = TRUE,
    omegaScale = 1000,
)
}
