% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\docType{class}
\name{ExperimentTranscriptionRates-class}
\alias{ExperimentTranscriptionRates-class}
\alias{plotRates,ExperimentTranscriptionRates-method}
\title{Constructor for ExperimentTranscriptionRates object}
\usage{
\S4method{plotRates}{ExperimentTranscriptionRates}(
  object,
  type = "scatter",
  rateType = "betaAdp",
  file = NULL,
  width = 8,
  height = 6,
  dpi = 300,
  ...
)
}
\description{
Class \code{ExperimentTranscriptionRates} has read counts, pause and gene
body genomic region coordinates, steric hindrance and omega scale factor
parameters used to estimate the transcription rates as well as the estimated
rates
}
\section{Slots}{

\describe{
\item{\code{counts}}{a \code{data.frame} with five columns gene_id,
summarizedPauseCounts, pauseLength, summarizedGbCounts, gbLength}

\item{\code{bigwigPlus}}{a path to bigwig for plus strand recording PRO-seq read
counts. This can be generated with the proseq2.0 pipeline.}

\item{\code{bigwigMinus}}{a path to bigwig for minus strand recording PRO-seq read
counts. This can be generated with the proseq2.0 pipeline.}

\item{\code{pauseRegions}}{a \code{\link[GenomicRanges]{GRanges-class}} that holds
all the pause region coordinates for every gene to handle read counting in
this region}

\item{\code{geneBodyRegions}}{a \code{\link[GenomicRanges]{GRanges-class}}
that holds all the gene body region coordinates for every gene to handle
read counting in this region}

\item{\code{geneNameColumn}}{a string for the gene name column in the GRanges}

\item{\code{stericHindrance}}{a logical value representing whether landing-pad
occupancy was inferred when estimating the rates. Landing pad occupancy
represents the probability that the landing pad required for a new
initiation event is already occupied by an RNAP If TRUE, the omegaScale
slot must be set to a numeric value greater than 0.}

\item{\code{omegaScale}}{a numeric for the scale factor used to calculate omega.
Omega represents the effective initiation rate, the initiation rate after a
portion of initiation events are blocked by RNAPs in the pause region due to
steric hindrance. Scale factors calibrate omega based on prior knowledge}

\item{\code{data}}{A \code{tbl_df} containing the estimated transcription rates:
\describe{
\item{\eqn{\chi}}{Numeric. Maximum likelihood estimate of the average read
depth in the gene body region.}
\item{\eqn{\beta_{org}}}{Numeric. Maximum likelihood estimate of average
read depth in pause region without varying pause sites}
\item{\eqn{\beta_{adp}}}{Numeric. Maximum likelihood estimate of average
read depth in pause region with varying pause sites}
\item{fkMean}{Numeric. Mean position of pause sites}
\item{fkVar}{Numeric. Variance of pause sites}
\item{\eqn{\phi}}{Numeric. Landing-pad occupancy estimates representing
probability of RNAP occupying the landing pad required for a new initiation
event}
\item{\eqn{\omega_{\zeta}}}{Numeric. Effective initiation rate, considering
steric hindrance}
\item{\eqn{\beta_{\zeta}}}{Numeric. Pause-escape rate}
\item{\eqn{\alpha_{\zeta}}}{Numeric. Potential initiation rate}
}}
}}

