% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\docType{class}
\name{experimentTranscriptionRates-class}
\alias{experimentTranscriptionRates-class}
\alias{plotRates,experimentTranscriptionRates-method}
\alias{rates}
\alias{rates,experimentTranscriptionRates-method}
\alias{counts}
\alias{counts,experimentTranscriptionRates-method}
\alias{pauseRegions}
\alias{pauseRegions,experimentTranscriptionRates-method}
\alias{geneBodyRegions}
\alias{geneBodyRegions,experimentTranscriptionRates-method}
\alias{geneNameColumn}
\alias{geneNameColumn,experimentTranscriptionRates-method}
\alias{omegaScale}
\alias{omegaScale,experimentTranscriptionRates-method}
\alias{stericHindrance,experimentTranscriptionRates-method}
\title{Class experimentTranscriptionRates}
\usage{
\S4method{plotRates}{experimentTranscriptionRates}(
  object,
  type = "scatter",
  rateType = "betaAdp",
  file = NULL,
  width = 8,
  height = 6,
  dpi = 300,
  ...
)

rates(object)

\S4method{rates}{experimentTranscriptionRates}(object)

counts(object)

\S4method{counts}{experimentTranscriptionRates}(object)

pauseRegions(object)

\S4method{pauseRegions}{experimentTranscriptionRates}(object)

geneBodyRegions(object)

\S4method{geneBodyRegions}{experimentTranscriptionRates}(object)

geneNameColumn(object)

\S4method{geneNameColumn}{experimentTranscriptionRates}(object)

omegaScale(object)

\S4method{omegaScale}{experimentTranscriptionRates}(object)

\S4method{stericHindrance}{experimentTranscriptionRates}(object)
}
\description{
Class \code{experimentTranscriptionRates} has read counts, pause and gene
body genomic region coordinates, steric hindrance and omega scale factor
parameters used to estimate the transcription rates
}
\section{Slots}{

\describe{
\item{\code{counts}}{a \code{data.frame} with five columns gene_id,
summarizedPauseCounts, pauseLength, summarizedGbCounts, gbLength}

\item{\code{bigwigPlus}}{a path to bigwig for plus strand}

\item{\code{bigwigMinus}}{a path to bigwig for minus strand}

\item{\code{pauseRegions}}{a \code{\link[GenomicRanges]{GRanges-class}} that holds
all the pause region coordinates}

\item{\code{geneBodyRegions}}{a \code{\link[GenomicRanges]{GRanges-class}}
that holds all the gene body region coordinates}

\item{\code{geneNameColumn}}{a string for the gene name column in the GRanges}

\item{\code{stericHindrance}}{a logical value representing whether landing-pad
occupancy was inferred when estimating the rates}

\item{\code{omegaScale}}{a numeric for the scale factor used to calculate omega}

\item{\code{rates}}{a \code{tbl_df} containing estimated transcription rates such as
chi estimates, betaOrg estimates from the initial model, betaAdp estimates
from the model with varying pause sites, fkMean giving the mean position
of pause sites, fkVar for variance of pause sites, phi estimates for
landing-pad occupancy, omegaZeta for the effective initiation rate,
betaZeta for the pause-escape rate, alphaZeta for the potential initiation
rate, and likelihoods}
}}

