% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{exportRatesToCSV}
\alias{exportRatesToCSV}
\title{Export rates to CSV}
\usage{
exportRatesToCSV(object, file = "experiment_rates.csv")
}
\arguments{
\item{object}{An experimentTranscriptionRates object}

\item{file}{Path to output CSV file. Defaults to "experiment_rates.csv"}
}
\value{
Outputs a CSV file with the rates
}
\description{
Export rates to CSV
}
\examples{
# Create an experimentTranscriptionRates object
expRates <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)

# Export rates to CSV
exportRatesToCSV(expRates, file = "experiment_rates.csv")
}
