% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{exportRatesToCSV}
\alias{exportRatesToCSV}
\title{Export rates to CSV}
\usage{
exportRatesToCSV(object, file)
}
\arguments{
\item{object}{An experiment_transcription_rates object}

\item{file}{Path to output CSV file}
}
\description{
Export rates to CSV
}
