% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{getRegions}
\alias{getRegions}
\title{Get genomic regions}
\usage{
getRegions(object, type)
}
\arguments{
\item{object}{An experiment_transcription_rates object}

\item{type}{Either "pause" or "gene_body" to specify which regions to return}
}
\value{
A GRanges object containing the specified regions
}
\description{
Get genomic regions
}
