% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_rates_lrts-class.R
\name{likelihoodRatioTest}
\alias{likelihoodRatioTest}
\title{likelihoodRatioTest}
\usage{
likelihoodRatioTest(expData1, expData2, sc)
}
\arguments{
\item{expData1}{an \code{\link{experimentTranscriptionRates-class}} object}

\item{expData2}{an \code{\link{experimentTranscriptionRates-class}} object}

\item{sc}{path to a csv file containing scale factors based on total or
spike-in reads}
}
\value{
a \code{\link{likelihoodRatioTest-class}} object
}
\description{
DESCRIPTION
}
