% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{parameters}
\alias{parameters}
\alias{parameters,simulatePolymerase-method}
\title{Get all simulation parameters}
\usage{
parameters(object)

\S4method{parameters}{simulatePolymerase}(object)
}
\arguments{
\item{object}{A simulatePolymerase object}
}
\value{
A list containing all simulation parameters including k, ksd, kMin, kMax,
geneLen, alpha, beta, zeta, zetaSd, zetaMin, zetaMax, zetaVec, cellNum,
polSize, addSpace, time, and stepsToRecord
}
\description{
Get all simulation parameters
}
