% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{plotPauseSites}
\alias{plotPauseSites}
\title{Plot pause site distribution}
\usage{
plotPauseSites(object, file = NULL, width = 8, height = 6)
}
\arguments{
\item{object}{A SimulatePolymerase object}

\item{file}{Optional file path to save the plot}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\value{
A ggplot object showing the distribution of pause sites
}
\description{
Plot pause site distribution
}
\examples{
# Create a SimulatePolymerase object
sim <- SimulatePolymerase(
    k=50, ksd=25, kMin=17, kMax=200, geneLen=1950,
    alpha=1, beta=1, zeta=2000, zetaSd=1000, zetaMin=1500, zetaMax=2500,
    zetaVec=NULL, cellNum=1000, polSize=33, addSpace=17, time=1, 
    stepsToRecord=1) 
# Plot pause site distribution
plotPauseSites(sim)
}
