% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{plotRates}
\alias{plotRates}
\title{Plot transcription rates}
\usage{
plotRates(object, type = "scatter", rate_type = "beta_adp", ...)
}
\arguments{
\item{object}{An experiment_transcription_rates object}

\item{type}{Type of plot to create ("scatter", "histogram", or "density")}

\item{rate_type}{Which rate to plot ("beta_org", "beta_adp", "chi", etc.)}

\item{...}{Additional arguments passed to the plotting function}
}
\value{
A ggplot object
}
\description{
Plot transcription rates
}
