% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-transcription-rates-class.R
\name{rates}
\alias{rates}
\title{Accessor for estimated rates}
\usage{
rates(object)
}
\arguments{
\item{object}{A TranscriptionRates object}
}
\value{
A tibble containing the estimated rates
}
\description{
Generic accessor for the estimated rates from any TranscriptionRates object
}
\examples{
# Works with both ExperimentTranscriptionRates and SimulationTranscriptionRates
rates(expRates)
rates(simRates)
}
