% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{sampleGeneBodyAvgReadDensity}
\alias{sampleGeneBodyAvgReadDensity}
\title{Sample read counts from a simulatePolymerase object}
\usage{
sampleGeneBodyAvgReadDensity(object, readDensity = 0.0489)
}
\arguments{
\item{object}{A simulatePolymerase object}

\item{readDensity}{A numeric value for the read density within gene body in
\emph{Dukler et al.} (2017) for genes with median expression (i.e., 0.0489).}
}
\value{
The average read density within gene body
}
\description{
Sample read counts from a simulatePolymerase object
}
\examples{
# Create a simulatePolymerase object
sim <- simulatePolymerase(
    k=50, ksd=25, kMin=17, kMax=200, geneLen=1950,
    alpha=1, beta=1, zeta=2000, zetaSd=1000, zetaMin=1500, zetaMax=2500,
    zetaVec=NULL, cellNum=1000, polSize=33, addSpace=17, time=1, 
    stepsToRecord=1) 
# Sample average read density within gene body
avgReadDensity <- sampleGeneBodyAvgReadDensity(sim)
# Print the average read density within gene body
print(avgReadDensity)
}
