% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{saveDataFrames}
\alias{saveDataFrames}
\title{Save all data frames to CSV files}
\usage{
saveDataFrames(object, dir = "results")
}
\arguments{
\item{object}{A simulatePolymerase object}

\item{dir}{Directory to save the files (default: "results")}
}
\value{
Outputs a CSV file with the data frames
}
\description{
Save all data frames to CSV files
}
\examples{
# Create a simulatePolymerase object
sim <- simulatePolymerase(
    k=50, ksd=25, kMin=17, kMax=200, geneLen=1950,
    alpha=1, beta=1, zeta=2000, zetaSd=1000, zetaMin=1500, zetaMax=2500,
    zetaVec=NULL, cellNum=1000, polSize=33, addSpace=17, time=1, 
    stepsToRecord=1)
# Save data frames
saveDataFrames(sim)
}
