% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\name{savePlots}
\alias{savePlots}
\title{Save all plots to files}
\usage{
savePlots(object, dir = "results", width = 8, height = 6)
}
\arguments{
\item{object}{A simulatePolymerase object}

\item{dir}{Directory to save the plots (default: "results")}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}
}
\value{
Outputs directory where plots were saved to pdf files
}
\description{
Save all plots to files
}
