% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_transcription_rates-class.R
\name{show,experimentTranscriptionRates-method}
\alias{show,experimentTranscriptionRates-method}
\title{Show method for experimentTranscriptionRates objects}
\usage{
\S4method{show}{experimentTranscriptionRates}(object)
}
\arguments{
\item{object}{An experimentTranscriptionRates object}
}
\value{
NULL (invisibly)
}
\description{
Show method for experimentTranscriptionRates objects
}
\examples{
# Create an experimentTranscriptionRates object
expRates <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "geneId"
)

# Show the object
show(expRates)
}
