% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polymerase-class.R
\docType{class}
\name{simulate_polymerase-class}
\alias{simulate_polymerase-class}
\alias{pause_sites}
\alias{probability_vector}
\alias{combined_cells_data}
\alias{position_matrix}
\alias{simulate_polymerase}
\alias{show,simulate_polymerase-method}
\alias{plot_probability}
\alias{plot_combined_cells}
\alias{plot_avg_read_density}
\alias{simulate_read_counts}
\alias{simulate_avg_read_density}
\alias{plot_probability,simulate_polymerase-method}
\alias{plot_combined_cells,simulate_polymerase-method}
\alias{plot_avg_read_density,simulate_polymerase-method}
\alias{simulate_read_counts,simulate_polymerase-method}
\alias{simulate_avg_read_density,simulate_polymerase-method}
\alias{k}
\alias{ksd}
\alias{k_min}
\alias{k_max}
\alias{gene_len}
\alias{alpha}
\alias{beta}
\alias{zeta}
\alias{zeta_sd}
\alias{zeta_min}
\alias{zeta_max}
\alias{cell_num}
\alias{pol_size}
\alias{add_space}
\alias{time}
\alias{steps_to_record}
\alias{read_counts}
\alias{avg_read_density}
\title{Class simulate_polymerase}
\usage{
pause_sites(object)

probability_vector(object)

combined_cells_data(object)

position_matrix(object)

simulate_polymerase(
  k,
  ksd,
  k_min,
  k_max,
  gene_len,
  alpha,
  beta,
  zeta,
  zeta_sd,
  zeta_min,
  zeta_max,
  cell_num,
  pol_size,
  add_space,
  time,
  steps_to_record
)

\S4method{show}{simulate_polymerase}(object)

plot_probability(object)

plot_combined_cells(object)

plot_avg_read_density(object)

simulate_read_counts(object)

simulate_avg_read_density(object)

\S4method{plot_probability}{simulate_polymerase}(object)

\S4method{plot_combined_cells}{simulate_polymerase}(object)

\S4method{plot_avg_read_density}{simulate_polymerase}(object)

\S4method{simulate_read_counts}{simulate_polymerase}(object)

\S4method{simulate_avg_read_density}{simulate_polymerase}(object)

k(object)

ksd(object)

k_min(object)

k_max(object)

gene_len(object)

alpha(object)

beta(object)

zeta(object)

zeta_sd(object)

zeta_min(object)

zeta_max(object)

cell_num(object)

pol_size(object)

add_space(object)

time(object)

steps_to_record(object)

read_counts(object)

avg_read_density(object)
}
\description{
Class \code{simulate_polymerase} tracks movement of RNAPs along the DNA
templates of a large number of cells. It accepts several key user-specified
parameters, including the initiation rate, pause-escape rate, a constant or
variable elongation rate, the mean and variance of pause sites across cells,
as well as the center-to-center spacing constraint between RNAPs, the number
of cells being simulated, the gene length, and the total time of
transcription. The simulator simply allows each RNAP to move forward or not,
in time slices of 1e-4 minutes, according to the specified position-specific
rate parameters. It assumes that at most one movement of each RNAP can occur
per time slice. The simulator monitors for collisions between adjacent
RNAPs, prohibiting one RNAP to advance if it is at the boundary of the
allowable distance from the next. After running for specified time, it
outputs the position of RNAPs for the last specified number of steps.
}
\section{Slots}{

\describe{
\item{\code{k}}{an integer value for the mean of pause sites across cells.}

\item{\code{ksd}}{a numeric value for the standard deviation of pause sites across
cells.}

\item{\code{k_min}}{an integer value for the upper bound of pause sites allowed.}

\item{\code{k_max}}{an integer value for the lower bound of pause sites allowed.}

\item{\code{gene_len}}{an integer value for the length of the gene.}

\item{\code{alpha}}{a numeric value for the initiation rate.}

\item{\code{beta}}{a numeric value for the pause release rate.}

\item{\code{zeta}}{a numeric value for the mean elongation rate across sites.}

\item{\code{zeta_sd}}{a numeric value for the standard deviation of pause sites
across sites.}

\item{\code{zeta_min}}{a numeric value for the minimum elongation rate.}

\item{\code{zeta_max}}{a numeric value for the maximum elongation rate.}

\item{\code{cell_num}}{an integer value for the number of cells to simulate.}

\item{\code{pol_size}}{an integer value for the polymerase II size.}

\item{\code{add_space}}{an integer value for the additional space in addition to
RNAP size.}

\item{\code{time}}{a numeric value for the time to simulate.}

\item{\code{steps_to_record}}{an integer value for the number of steps to record in
position matrix.}

\item{\code{pause_sites}}{a numeric vector of pause sites}

\item{\code{probability_vector}}{a numeric vector}

\item{\code{combined_cells_data}}{an integer vector}

\item{\code{position_matrix}}{a matrix of position of polymerase}

\item{\code{read_counts}}{a numeric vector for read counts}
}}

