% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\docType{class}
\name{simulation_transcription_rates-class}
\alias{simulation_transcription_rates-class}
\title{Class simulation_transcription_rates}
\description{
Class \code{simulation_transcription_rates} has
}
