% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_transcription_rates-class.R
\docType{class}
\name{simulationTranscriptionRates-class}
\alias{simulationTranscriptionRates-class}
\alias{show,simulationTranscriptionRates-method}
\alias{simpol}
\alias{stericHindrance,simulationTranscriptionRates-method}
\alias{trial}
\alias{chi}
\alias{betaOrg}
\alias{betaAdp}
\alias{phi}
\alias{fk}
\alias{fkMean}
\alias{fkVar}
\alias{flag}
\alias{rnapN}
\title{Class simulationTranscriptionRates}
\usage{
\S4method{show}{simulationTranscriptionRates}(object)

simpol(object)

\S4method{stericHindrance}{simulationTranscriptionRates}(object)

trial(object)

chi(object)

betaOrg(object)

betaAdp(object)

phi(object)

fk(object)

fkMean(object)

fkVar(object)

flag(object)

rnapN(object)
}
\description{
Class \code{simulationTranscriptionRates} estimates the transcription
rates, such as initiation, pause-release rates and landing pad occupancy,
from simulated data, such as nascent RNA sequencing read counts and genomic
coordinates simulated from polymerase movement over a specified time period
calculated by simulator, and contructs an object that holds these rates.
}
