% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-transcription-rates-class.R
\name{stericHindrance}
\alias{stericHindrance}
\title{Accessor for steric hindrance flag}
\usage{
stericHindrance(object)
}
\arguments{
\item{object}{A TranscriptionRates object}
}
\value{
A logical value indicating whether steric hindrance was modeled
}
\description{
Generic accessor for the steric hindrance flag from any TranscriptionRates
object
}
\examples{
# Create an ExperimentTranscriptionRates object
load("inst/extdata/granges_for_read_counting_chr21_subset.RData")
expRates <- estimateTranscriptionRates(
    "inst/extdata/PROseq-K562-vihervaara-control-SE_plus_chr21_subset.bw",
    bigwigMinus = 
     "inst/extdata/PROseq-K562-vihervaara-control-SE_minus_chr21_subset.bw",
    pauseRegions = bw_pause_21_subset,
    geneBodyRegions = bw_gene_body_21_subset,
    stericHindrance = TRUE,
    omegaScale = 1000,
)
stericHindrance(expRates)

}
