% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{stericHindrance}
\alias{stericHindrance}
\title{Generic function for stericHindrance}
\usage{
stericHindrance(object)
}
\arguments{
\item{object}{An object}
}
\value{
The stericHindrance value
}
\description{
Generic function for stericHindrance
}
\examples{
expRates <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
stericHindrance(expRates)
}
