% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-transcription-rates-class.R
\name{stericHindrance}
\alias{stericHindrance}
\title{Accessor for steric hindrance flag}
\usage{
stericHindrance(object)
}
\arguments{
\item{object}{A TranscriptionRates object}
}
\value{
A logical value indicating whether steric hindrance was modeled
}
\description{
Generic accessor for the steric hindrance flag from any TranscriptionRates object
}
\examples{
# Works with both ExperimentTranscriptionRates and SimulationTranscriptionRates
stericHindrance(expRates)
stericHindrance(simRates)
}
