% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{stericHindrance}
\alias{stericHindrance}
\title{Generic function for stericHindrance}
\usage{
stericHindrance(object)
}
\arguments{
\item{object}{An object}
}
\value{
The stericHindrance value
}
\description{
Generic function for stericHindrance
}
