% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-transcription-rates-class.R
\docType{class}
\name{TranscriptionRates-class}
\alias{TranscriptionRates-class}
\title{Base class for TranscriptionRates objects}
\description{
Virtual base class that defines the common interface for transcription rate
objects. Both \code{ExperimentTranscriptionRates} and
\code{SimulationTranscriptionRates} inherit from this class.
}
\section{Slots}{

\describe{
\item{\code{rates}}{a \code{\link[tibble]{tbl_df}} containing the estimated rates}

\item{\code{stericHindrance}}{a logical value indicating whether steric hindrance
was modeled}
}}

