% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_rates_lrts-class.R
\docType{class}
\name{transcriptionRatesLrts-class}
\alias{transcriptionRatesLrts-class}
\alias{expData1}
\alias{expData2}
\alias{sc}
\title{Class transcriptionRatesLrts for comparing aspects of transcriptional
dynamics of the same TU under different conditions using Likelihood Ratio
Test Statistics}
\usage{
expData1(object)

expData2(object)

sc(object)
}
\description{
Class \code{transcriptionRatesLrts}
}
\examples{
expData1 <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
expData2 <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
lrts <- likelihoodRatioTest(expData1, expData2, sc)
expData1(lrts)
expData1 <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
expData2 <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
lrts <- likelihoodRatioTest(expData1, expData2, sc)
expData2(lrts)
expData1 <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
expData2 <- estimateExperimentTranscriptionRates(
    bigwigPlus = "path/to/plus.bw",
    bigwigMinus = "path/to/minus.bw",
    pauseRegions = GRanges("chr1:1-1000"),
    geneBodyRegions = GRanges("chr1:1-2000"),
    geneNameColumn = "gene_id"
)
lrts <- likelihoodRatioTest(expData1, expData2, sc)
sc(lrts)
}
